requirements.txt
# LLMbreaker 依赖
# 核心依赖
requests>=2.28.0
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# 数据处理
pandas>=2.0.0
openpyxl>=3.1.0

# 进度显示
tqdm>=4.65.0

# 数据库
pymysql>=1.1.0

# 配置文件
pyyaml>=6.0

# IntentLang 融合模块依赖
# Pydantic - 强类型输出验证
pydantic>=2.0.0

# 并发控制
# concurrent-futures 已内置于 Python 3.x

# 可选依赖（用于高级功能）
# httpx>=0.24.0  # 异步HTTP客户端
# aiohttp>=3.8.0  # 异步HTTP
# redis>=4.5.0  # 分布式缓存
